# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Time-Entry APIs",
    "description": "Collection of APIs to manage the time entry module.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    }
  },
  "host": "localhost:8080",
  "basePath": "/api/exp",
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"."
    }
  },
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/auth/v1/register": {
      "post": {
        "description": "Register user with email and password",
        "summary": "RegisterUser",
        "tags": ["Authentication"],
        "operationId": "RegisterUser",
        "deprecated": false,
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/RegisterUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Response",
            "schema": {
              "type": "object",
              "properties": {
                "isError": {
                  "type": "boolean",
                  "example": false
                },
                "message": {
                  "type": "string",
                  "example": "Successfully registered!"
                }
              }
            }
          },
          "400": {
            "description": "Error Response",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": []
      }
    },
...
  },
  "definitions": {
...
    "RegisterUserRequest": {
      "title": "RegisterUserRequest",
      "example": {
        "userName": "user2@test.com",
        "password": "password@123"
      },
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": ["userName", "password"]
    },
...
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "isError": {
          "type": "boolean",
          "example": true
        },
        "message": {
          "type": "string",
          "example": "Failed"
        }
      }
    },
...
  },
  "security": [],
...
}